add: 약관 및 정책
