+++ src/main/java/umc/meme/shop/domain/mypage/controller/MyPageController.java
@@ -3,6 +3,7 @@ package umc.meme.shop.domain.mypage.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
+import umc.meme.shop.domain.mypage.service.MypageService;
 import umc.meme.shop.global.SuccessStatus;
 import umc.meme.shop.global.response.ApiResponse;
 
@@ -11,10 +12,29 @@ import umc.meme.shop.global.response.ApiResponse;
 @RequestMapping("/api/v1/mypage")
 public class MyPageController {
 
-    @Operation(summary = "마이페이지", description = "마이페이지를 조회하는 API입니다.")
-    @GetMapping("/{userId}")
-    public ApiResponse getMypage(@PathVariable Long userId){
-        return ApiResponse.SuccessResponse(SuccessStatus.MYPAGE_GET, "");
+    private final MypageService mypageService;
+
+    /**
+     * 마이페이지 조회는
+     * 모델/아티스트 에서 처리
+     */
+
+//    @Operation(summary = "마이페이지", description = "마이페이지를 조회하는 API입니다.")
+//    @GetMapping("/{userId}")
+//    public ApiResponse getMypage(@PathVariable Long userId){
+//        return ApiResponse.SuccessResponse(SuccessStatus.MYPAGE_GET, "");
+//    }
+
+    @Operation(summary = "모델 마이페이지 조회")
+    @GetMapping("/{modelId}/profile/model")
+    public ApiResponse updateModelProfile (@PathVariable Long modelId){
+        return ApiResponse.SuccessResponse(SuccessStatus.MYPAGE_GET, mypageService.getModelProfile(modelId));
+    }
+
+    @Operation(summary = "아티스트 마이페이지 조회")
+    @GetMapping("/{artistId}/profile/artist")
+    public ApiResponse getProfile(@PathVariable Long artistId){
+        return ApiResponse.SuccessResponse(SuccessStatus.MYPAGE_GET, mypageService.getArtistProfile(artistId));
     }
 
     @Operation(summary = "내 정보 조회", description = "본인 정보를 조회하는 API입니다.")
@@ -27,7 +47,7 @@ public class MyPageController {
     @Operation(summary = "약관 및 정책", description = "약관 및 정책을 조회하는 API입니다.")
     @GetMapping("/tos")
     public ApiResponse tos(){
-        return ApiResponse.SuccessResponse(SuccessStatus.TOS_GET, "");
+        return ApiResponse.SuccessResponse(SuccessStatus.TOS_GET, mypageService.getTos());
     }
 
     @Operation(summary = "문의하기", description = "문의하기 API입니다.")
