+++ src/main/java/umc/meme/shop/domain/mypage/dto/response/MypageTosDto.java
@@ -1,0 +1,10 @@
+package umc.meme.shop.domain.mypage.dto.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class MypageTosDto {
+    private String tos;
+}
